+++ components/index.ts
@@ -1,4 +1,8 @@
-export * from './Feature';
-export * from './Grid';
+export * from './Carousel';
+export * from './Downloads';
+export * from './Footer';
 export * from './Page';
-export * from './Teaser';
+export * from './StaticImage';
+export * from './Header';
+export * from './Layout';
+