+++ config.ts
@@ -1,0 +1,6 @@
+const config = {
+	// TODO: Add Google Analytics ID
+	GTM_ID: 'GTM-XXXXXXX',
+};
+
+export default config;