+++ pages/_document.tsx
@@ -1,0 +1,52 @@
+import Document, { Head, Html, Main, NextScript } from 'next/document';
+import siteConfig from '../config';
+
+class MyDocument extends Document {
+	render() {
+		return (
+			<Html lang="en">
+				<Head>
+					<meta name="description" content="Furniture Group" />
+					{/* <!-- TODO enable when going live --> */}
+					{/* <script
+						dangerouslySetInnerHTML={{
+							__html: `(function (w, d, s, l, i) {
+								w[l] = w[l] || [];
+								w[l].push({
+									'gtm.start': new Date().getTime(),
+									event: 'gtm.js',
+								});
+								var f = d.getElementsByTagName(s)[0],
+									j = d.createElement(s),
+									dl = l != 'dataLayer' ? '&l=' + l : '';
+								j.async = true;
+								j.src = 'https://www.googletagmanager.com/gtm.js?id=' + i + dl;
+								f.parentNode.insertBefore(j, f);
+							})(window, document, 'script', 'dataLayer', 'GTM-5BPRQM');`,
+						}}
+					/> */}
+				</Head>
+
+				<body>
+					<Main></Main>
+					<NextScript />
+					{/* TODO: enable when going live */}
+					{/* {siteConfig.GTM_ID && (
+						<noscript
+							dangerouslySetInnerHTML={{
+								__html: `<iframe
+								src="https://www.googletagmanager.com/ns.html?id=GTM-${siteConfig.GTM_ID}"
+								height="0"
+								width="0"
+								style="display: none; visibility: hidden"
+							></iframe>`,
+							}}
+						/>
+					)} */}
+				</body>
+			</Html>
+		);
+	}
+}
+
+export default MyDocument;