+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] -05/14/2023
+
+### Added
+-   Add landing page for neocon (header, hero banner, image carousel, downloads, footer)
+-   Add gtm for tracking
+-   Add local and google fonts
+-   Add skip to main navigation
+
+### Changed
+
+### Fixed
+
+### Removed
+
+-   Feature, Grid and Teaser component
+
 ## [1.0.2] -12/16/2022
 
 ### Fixed